TMDB movie dataset analysis
This Jupyter Notebook looks into a TMDB movies data set. In the analysis, I tried to find any interesting insights that could be an important factor in determining whether a movie will be highly popular or not, and whether it will be a top grossing movie in terms of revenue or not. All the work is described in details and a conclusion is provided at the end of the notebook.